+++ llm/gemini.go
@@ -1,4 +1,4 @@
-package utils
+package llm
 
 import (
 	"context"
